rename: patterns.ts -> lgPatterns.ts
