+++ Composer/packages/lib/shared/src/lgUtils/lgPatterns.ts
